+++ R/mvgls.dfa.r
@@ -232,4 +232,135 @@ print.mvgls.dfa <- function(x, ...){
   cat("\n")
 }
 
+# ------------------------------------------------------------------------- #
+# predict.mvgls.dfa                                                         #
+# options: object, newdata, prior, ...                                      #
+#                                                                           #
+# ------------------------------------------------------------------------- #
+predict.mvgls.dfa <- function(object, newdata, prior = object$prior, ...){
+    
+    args <- list(...)
+    if(is.null(args[["tree"]])) tree <- NULL else tree <- args$tree
+    # Checks that the model is based on dummy contrasts
+    contrasts_types <- object$fit$contrasts[attr(object$fit$terms, "term.labels")]
+    if(contrasts_types[object$term]!="contr.treatment") warning("Model assumed that the contrasts for the term of interest is of type \"contr.treatment\"")
+    # Check if there's extra terms in the model
+    if(sum(object$classid)!=ncol(object$fit$variables$X)) stop("The DFA function is not currently working for multi-predictors models") # FIXME
+    
+    # Retrieve coefficients
+    B <- coef(object$fit)
+    index_B <- (1:nrow(B))[object$classid]
+    
+    # FIXME: works for treatment contrasts. To be generalized
+    if(any(object$fit$dims$assign==0)){
+        B[object$fit$dims$assign==object$term,] <-  sweep(B[object$fit$dims$assign==object$term,,drop=FALSE], 2, B[object$fit$dims$assign==0,], "+")
+    }
+    
+    # log-sum-exp trick to avoid over/under flow
+    logsumexp <- function(v) max(v) + log(sum(exp(v - max(v))))
+    
+    if(is.null(tree)){
+        #if no datasets provided, classification rule is applied to the training dataset > see 9.12 in Rencher 2002
+        # Bayes classifier for group "g" is d(x) = t(x)S^-1u - 0.5t(u)S^-1u + log(P(g)) => can be used to precompute quantities for multiple values to predict
+        if(missing(newdata)){
+            prediction <- sapply(index_B, function(i){
+                SB <- object$fit$sigma$P%*%B[i,]
+                const_prior <- -0.5*t(B[i,])%*%SB + log(prior[i])
+                sapply(1:nrow(object$fit$variable$Y), function(x){
+                    t(object$fit$variable$Y[x,])%*%SB + const_prior
+                })
+            })
+            
+        }else{
+            prediction <- sapply(index_B, function(i){
+                SB <- object$fit$sigma$P%*%B[i,]
+                const_prior <- -0.5*t(B[i,])%*%SB + log(prior[i])
+                sapply(1:nrow(newdata), function(x){
+                    t(newdata[x,])%*%SB + const_prior
+                })
+            })
+            if(!is.matrix(prediction))  prediction <- matrix(prediction, nrow=nrow(newdata))
+        }
+        
+        # Compute the posterior and classification
+        
+        posterior <- exp(sweep(prediction, 1, apply(prediction,1,logsumexp), "-")) # FIXME to check
+        classif <- apply(posterior, 1, which.max)
+        
+        # rename
+        names_variables <- attr(object$fit$variables$X,"dimnames")[[2]][object$classid]
+        colnames(posterior) <- names_variables
+        if(missing(newdata)) rownames(posterior) = rownames(object$fit$variable$Y) else rownames(posterior) = rownames(newdata)
+        if(missing(newdata)) names(classif) = rownames(object$fit$variable$Y) else names(classif) = rownames(newdata)
+        names(prior) <- names_variables
+        
+    }else{
+        
+        # checks
+        if(!inherits(tree, "phylo")) stop("You must provide a tree object of class \"phylo\"")
+        if(!is.data.frame(newdata) & !is.matrix(newdata)) stop("the \"newdata\" should be a data.frame object with column names matching predictors names, and row names matching names in the tree ")
+        # prep.
+        predicted_names <- rownames(newdata)
+        rcov <- .resid_cov_phylo(tree, object$fit, predicted_names)
+        X1 <- matrix(1,ncol=1,nrow=nrow(newdata))
+        # modify the design matrix FIXME, is it necessary?
+        if(any(object$fit$dims$assign==0)){
+            grp <- as.factor(object$fit$variables$X[,object$classid]%*%(1:object$nclass))
+            Xn <- model.matrix(~grp+0)
+            resid <- object$fit$variable$Y - Xn%*%B
+        }else{
+            resid <- residuals(object$fit)
+        }
+        # we compute the residuals across all subjects
+        
+        # compute prediction scores
+        prediction <- sapply(1:nrow(newdata), function(x){
+            sapply(1:nrow(B), function(i){
+                bias <- rcov$w%*%solve(rcov$Vt)%*%resid[rcov$train,,drop=FALSE]
+                predicted <- X1%*%B[i,] + bias
+                -0.5*( t(as.numeric(newdata[x,] - predicted[x,]))%*%object$fit$sigma$P%*%as.numeric(newdata[x,] - predicted[x,])) + log(prior[i]) #eg eq 26 in Hastie et al. 1994 - PDA
+            })
+        })
+        
+        # Compute the posterior and classification
+        posterior <- t(exp(sweep(prediction, 2, apply(prediction, 2, logsumexp), "-"))) # FIXME to check
+        classif <- apply(posterior, 1, which.max)
+        
+        # rename
+        #colnames(posterior) <- attr(object$fit$variables$X,"dimnames")[[2]][object$classid]
+        rownames(posterior) <- predicted_names
+        # names(prior) <- attr(object$fit$variables$X,"dimnames")[[2]][object$classid]
+        names(classif) <- predicted_names
+    }
+    
+    
+    # results
+    if(missing(newdata)){
+        confusion <- table(classif, object$fit$variables$X[,object$classid]%*%(1:object$nclass) ) # FIXME provide the categorical variable instead?
+        rownames(confusion) = colnames(confusion) = names_variables
+        results <- list(class=classif, posterior=posterior, prior=prior, confusion=confusion, training=TRUE)
+    }else{
+        results <- list(class=classif, posterior=posterior, prior=prior, training=FALSE)
+    }
+    class(results) <- "mvgls.dfa.predict"
+    return(results)
+}
+
+# ------------------------------------------------------------------------- #
+# print.mvgls.dfa.predict                                                   #
+# options: x, ...                                                           #
+#                                                                           #
+# ------------------------------------------------------------------------- #
+print.mvgls.dfa.predict <- function(x, ...){
+    cat("\n")
+    message("-- DFA predictions --","\n")
+    cat("Prior probabilities of groups:","\n")
+    cat(signif(x$prior, digits=4),"\n")
+    cat("\n")
+    if(x$training){
+        error = signif(sum(diag(x$confusion))/length(x$class), digits = 3)
+        cat("Training misclassification error rate:",1-error,"\n")
+    }
+    cat("\n")
+}